requirements.txt
Flask==3.0.2
dnspython==2.5.0
requests==2.31.0